Library name is now ledmatrix4j
rename module, package and update doc accordingly
