--- ledmatrix/src/main/java/fr/craftinglabs/pi/matrix/io/MatrixIO.java
@@ -1,12 +1,0 @@
-package fr.craftinglabs.pi.matrix.io;
-
-public interface MatrixIO {
-
-    void incrementCounter();
-
-    void resetCounter();
-
-    void releaseShift() ;
-
-    void shiftLine(byte value);
-}
