+++ ledmatrix4j/src/main/java/fr/craftinglabs/pi/ledmatrix4j/Size.java
@@ -1,0 +1,20 @@
+package fr.craftinglabs.pi.ledmatrix4j;
+
+public class Size {
+
+    private final int length;
+    private final int height;
+
+    public Size(int height, int length) {
+        this.height = height;
+        this.length = length;
+    }
+    
+    public int getHeight() {
+        return height;
+    }
+    
+    public int getLength() {
+        return length;
+    }
+}
