+++ ledmatrix4j/src/main/java/fr/craftinglabs/pi/ledmatrix4j/font/Font.java
@@ -1,0 +1,16 @@
+package fr.craftinglabs.pi.ledmatrix4j.font;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class Font {
+    static final Map<Character, Glyph> glyphs = new HashMap<>();
+
+    static public void add(Glyph glyph) {
+        glyphs.put(glyph.getChar(), glyph);
+    }
+
+    static public Glyph glyphFor(Character character) {
+        return glyphs.get(character);
+    }
+}
