+++ ledmatrix4j/src/main/java/fr/craftinglabs/pi/ledmatrix4j/font/Font6.java
@@ -1,0 +1,638 @@
+package fr.craftinglabs.pi.ledmatrix4j.font;
+
+public class Font6 extends Font {
+
+    static final public Glyph A = new Glyph('A',
+            "011100",
+            "100010",
+            "100010",
+            "111110",
+            "100010",
+            "100010");
+
+    static final public Glyph B = new Glyph('B',
+            "111100",
+            "100010",
+            "100010",
+            "111100",
+            "100010",
+            "111110");
+
+    static final public Glyph C = new Glyph( 'C',
+            "111110",
+            "100000",
+            "100000",
+            "100000",
+            "100000",
+            "111110");
+
+
+    static final public Glyph D = new Glyph( 'D',
+            "111100",
+            "100010",
+            "100010",
+            "100010",
+            "100010",
+            "111100");
+
+    static final public Glyph E = new Glyph( 'E',
+            "111110",
+            "100000",
+            "100000",
+            "111100",
+            "100000",
+            "111110");
+
+    static final public Glyph F = new Glyph( 'F',
+            "111110",
+            "100000",
+            "100000",
+            "111100",
+            "100000",
+            "100000");
+
+    static final public Glyph G = new Glyph( 'G',
+            "011100",
+            "100010",
+            "100000",
+            "100110",
+            "100010",
+            "011100");
+
+    static final public Glyph H = new Glyph( 'H',
+            "100010",
+            "100010",
+            "111110",
+            "100010",
+            "100010",
+            "100010");
+
+    static final public Glyph I = new Glyph( 'I',
+            "111110",
+            "001000",
+            "001000",
+            "001000",
+            "001000",
+            "111110");
+
+    static final public Glyph J = new Glyph( 'J',
+            "001110",
+            "000100",
+            "000100",
+            "000100",
+            "100100",
+            "011000");
+
+    static final public Glyph M = new Glyph( 'M',
+            "100010",
+            "110110",
+            "101010",
+            "101010",
+            "100010",
+            "100010");
+
+    static final public Glyph N = new Glyph( 'N',
+            "100010",
+            "110010",
+            "101010",
+            "101010",
+            "100110",
+            "100010");
+
+    static final public Glyph L = new Glyph( 'L',
+            "100000",
+            "100000",
+            "100000",
+            "100000",
+            "100000",
+            "111110");
+
+    static final public Glyph O = new Glyph( 'O',
+            "011100",
+            "100010",
+            "100010",
+            "100010",
+            "100010",
+            "011100");
+
+    static final public Glyph P = new Glyph( 'P',
+            "111100",
+            "100010",
+            "100010",
+            "111100",
+            "100000",
+            "100000");
+
+    static final public Glyph Q = new Glyph( 'Q',
+            "011100",
+            "100010",
+            "101010",
+            "100110",
+            "011110",
+            "000010");
+
+    static final public Glyph R = new Glyph( 'R',
+            "111100",
+            "100010",
+            "100010",
+            "111100",
+            "100010",
+            "100010");
+
+    static final public Glyph S = new Glyph( 'S',
+            "011100",
+            "100010",
+            "011000",
+            "000100",
+            "100010",
+            "011100");
+
+    static final public Glyph K = new Glyph( 'K',
+            "100010",
+            "100100",
+            "111000",
+            "111000",
+            "100100",
+            "100010");
+
+    static final public Glyph T = new Glyph( 'T',
+            "111110",
+            "001000",
+            "001000",
+            "001000",
+            "001000",
+            "001000");
+
+    static final public Glyph U = new Glyph( 'U',
+            "100010",
+            "100010",
+            "100010",
+            "100010",
+            "100010",
+            "011100");
+
+    static final public Glyph V = new Glyph( 'V',
+            "100010",
+            "100010",
+            "100010",
+            "100010",
+            "010100",
+            "001000");
+
+    static final public Glyph W = new Glyph( 'W',
+            "100010",
+            "100010",
+            "101010",
+            "101010",
+            "101010",
+            "010100");
+
+    static final public Glyph X = new Glyph( 'X',
+            "100010",
+            "010100",
+            "001000",
+            "001000",
+            "010100",
+            "100010");
+
+    static final public Glyph Y = new Glyph( 'Y',
+            "100010",
+            "010100",
+            "001000",
+            "001000",
+            "001000",
+            "001000");
+
+    static final public Glyph Z = new Glyph( 'Z',
+            "111110",
+            "000010",
+            "001100",
+            "011000",
+            "100000",
+            "111110");
+
+    static final public Glyph a = new Glyph( 'a',
+            "000000",
+            "011100",
+            "000010",
+            "011110",
+            "100010",
+            "011110");
+
+    static final public Glyph b = new Glyph( 'b',
+            "100000",
+            "100000",
+            "101100",
+            "110010",
+            "100010",
+            "111100");
+
+    static final public Glyph c = new Glyph( 'c',
+            "000000",
+            "011100",
+            "100000",
+            "100000",
+            "100010",
+            "011100");
+
+    static final public Glyph d = new Glyph( 'd',
+            "000010",
+            "000010",
+            "011110",
+            "100010",
+            "100010",
+            "011110");
+
+    static final public Glyph e = new Glyph( 'e',
+            "000000",
+            "011100",
+            "100010",
+            "111110",
+            "100000",
+            "011100");
+
+    static final public Glyph f = new Glyph( 'f',
+            "001100",
+            "010010",
+            "010000",
+            "111000",
+            "010000",
+            "010000");
+
+    static final public Glyph g = new Glyph( 'g',
+            "000000",
+            "011110",
+            "100010",
+            "011110",
+            "000010",
+            "011100");
+
+    static final public Glyph h = new Glyph( 'h',
+            "100000",
+            "100000",
+            "101100",
+            "110010",
+            "100010",
+            "100010");
+
+    static final public Glyph i = new Glyph( 'i',
+            "001000",
+            "000000",
+            "011000",
+            "001000",
+            "001000",
+            "011110");
+
+    static final public Glyph j = new Glyph( 'j',
+            "000100",
+            "000000",
+            "001110",
+            "000100",
+            "100100",
+            "011000");
+
+    static final public Glyph k = new Glyph( 'k',
+            "10000",
+            "10010",
+            "10100",
+            "11000",
+            "10100",
+            "10010");
+
+    static final public Glyph l = new Glyph( 'l',
+            "11000",
+            "01000",
+            "01000",
+            "01000",
+            "01000",
+            "11110");
+
+    static final public Glyph m = new Glyph( 'm',
+            "000000",
+            "000000",
+            "110100",
+            "101010",
+            "101010",
+            "100010");
+
+    static final public Glyph n = new Glyph( 'n',
+            "000000",
+            "000000",
+            "101100",
+            "110010",
+            "100010",
+            "100010");
+
+    static final public Glyph o = new Glyph( 'o',
+            "000000",
+            "011100",
+            "100010",
+            "100010",
+            "100010",
+            "011100");
+
+    static final public Glyph p = new Glyph( 'p',
+            "000000",
+            "111100",
+            "100010",
+            "111100",
+            "100000",
+            "100000");
+
+    static final public Glyph q = new Glyph( 'q',
+            "000000",
+            "011010",
+            "100110",
+            "011110",
+            "000010",
+            "000010");
+
+    static final public Glyph r = new Glyph( 'r',
+            "000000",
+            "000000",
+            "101100",
+            "110010",
+            "100000",
+            "100000");
+
+    static final public Glyph s = new Glyph( 's',
+            "000000",
+            "011100",
+            "100000",
+            "011100",
+            "000010",
+            "111100");
+
+    static final public Glyph t = new Glyph( 't',
+            "010000",
+            "010000",
+            "111000",
+            "010000",
+            "010010",
+            "001100");
+
+    static final public Glyph u = new Glyph( 'u',
+            "000000",
+            "000000",
+            "100010",
+            "100010",
+            "100110",
+            "011010");
+
+    static final public Glyph v = new Glyph( 'v',
+            "000000",
+            "000000",
+            "100010",
+            "100010",
+            "010100",
+            "001000");
+
+    static final public Glyph w = new Glyph( 'w',
+            "000000",
+            "000000",
+            "100010",
+            "101010",
+            "101010",
+            "010100");
+
+    static final public Glyph x = new Glyph( 'x',
+            "000000",
+            "100010",
+            "010100",
+            "001000",
+            "010100",
+            "100010");
+
+    static final public Glyph y = new Glyph( 'y',
+            "000000",
+            "100010",
+            "100010",
+            "011110",
+            "000010",
+            "011100");
+
+    static final public Glyph z = new Glyph( 'z',
+            "000000",
+            "111110",
+            "000100",
+            "001000",
+            "010000",
+            "111110");
+
+    static final public Glyph SPACE = new Glyph( ' ',
+            "0000",
+            "0000",
+            "0000",
+            "0000",
+            "0000",
+            "0000");
+
+    static final public Glyph NUM0 = new Glyph( '0',
+            "011100",
+            "100110",
+            "101010",
+            "101010",
+            "110010",
+            "011100");
+
+    static final public Glyph NUM1 = new Glyph( '1',
+            "00100",
+            "01100",
+            "10100",
+            "00100",
+            "00100",
+            "01110");
+
+    static final public Glyph NUM2 = new Glyph( '2',
+            "011100",
+            "100010",
+            "000010",
+            "011100",
+            "100000",
+            "111110");
+
+    static final public Glyph NUM3 = new Glyph( '3',
+            "111100",
+            "000010",
+            "000010",
+            "011110",
+            "000010",
+            "111100");
+
+    static final public Glyph NUM4 = new Glyph( '4',
+            "100010",
+            "100010",
+            "100010",
+            "111110",
+            "000010",
+            "000010");
+
+    static final public Glyph NUM5 = new Glyph( '5',
+            "111110",
+            "100000",
+            "111100",
+            "000010",
+            "100010",
+            "011100");
+
+    static final public Glyph NUM6 = new Glyph( '6',
+            "111110",
+            "100000",
+            "111110",
+            "100010",
+            "100010",
+            "111110");
+
+    static final public Glyph NUM7 = new Glyph( '7',
+            "111110",
+            "000010",
+            "000010",
+            "011110",
+            "000010",
+            "000010");
+
+    static final public Glyph NUM8 = new Glyph( '8',
+            "111110",
+            "100010",
+            "111110",
+            "100010",
+            "100010",
+            "111110");
+
+    static final public Glyph NUM9 = new Glyph( '9',
+            "111110",
+            "100010",
+            "111110",
+            "000010",
+            "000010",
+            "111110");
+
+    static final public Glyph DIVIDE = new Glyph( '/',
+            "000010",
+            "000100",
+            "001000",
+            "001000",
+            "010000",
+            "100000");
+
+    static final public Glyph COLUMN = new Glyph( ':',
+            "0110",
+            "0110",
+            "0000",
+            "0000",
+            "0110",
+            "0110");
+
+    static final public Glyph DOT = new Glyph( '.',
+            "0000",
+            "0000",
+            "0000",
+            "0000",
+            "0110",
+            "0110");
+
+    static final public Glyph COMA = new Glyph( ',',
+            "00000",
+            "00000",
+            "00000",
+            "00110",
+            "00110",
+            "01100");
+
+    static final public Glyph DASH = new Glyph( '-',
+            "000000",
+            "000000",
+            "111110",
+            "111110",
+            "000000",
+            "000000");
+
+    static final public Glyph QUESTION = new Glyph( '?',
+            "011100",
+            "100010",
+            "000100",
+            "001000",
+            "000000",
+            "001000");
+
+    static final public Glyph MARK = new Glyph( '!',
+            "00100",
+            "01110",
+            "01110",
+            "00100",
+            "00000",
+            "00100");
+
+    static {
+        add(Font6.A);
+        add(Font6.B);
+        add(Font6.C);
+        add(Font6.D);
+        add(Font6.E);
+        add(Font6.F);
+        add(Font6.G);
+        add(Font6.H);
+        add(Font6.I);
+        add(Font6.J);
+        add(Font6.K);
+        add(Font6.L);
+        add(Font6.M);
+        add(Font6.N);
+        add(Font6.O);
+        add(Font6.P);
+        add(Font6.Q);
+        add(Font6.R);
+        add(Font6.S);
+        add(Font6.T);
+        add(Font6.U);
+        add(Font6.V);
+        add(Font6.W);
+        add(Font6.X);
+        add(Font6.Y);
+        add(Font6.Z);
+        add(Font6.a);
+        add(Font6.b);
+        add(Font6.c);
+        add(Font6.d);
+        add(Font6.e);
+        add(Font6.f);
+        add(Font6.g);
+        add(Font6.h);
+        add(Font6.i);
+        add(Font6.j);
+        add(Font6.k);
+        add(Font6.l);
+        add(Font6.m);
+        add(Font6.n);
+        add(Font6.o);
+        add(Font6.p);
+        add(Font6.q);
+        add(Font6.r);
+        add(Font6.s);
+        add(Font6.t);
+        add(Font6.u);
+        add(Font6.v);
+        add(Font6.w);
+        add(Font6.x);
+        add(Font6.y);
+        add(Font6.z);
+        add(Font6.NUM0);
+        add(Font6.NUM1);
+        add(Font6.NUM2);
+        add(Font6.NUM3);
+        add(Font6.NUM4);
+        add(Font6.NUM5);
+        add(Font6.NUM6);
+        add(Font6.NUM7);
+        add(Font6.NUM8);
+        add(Font6.NUM9);
+        add(Font6.DASH);
+        add(Font6.DIVIDE);
+        add(Font6.DOT);
+        add(Font6.COLUMN);
+        add(Font6.COMA);
+        add(Font6.SPACE);
+        add(Font6.QUESTION);
+        add(Font6.MARK);
+    }
+}
