+++ ledmatrix4j/src/main/java/fr/craftinglabs/pi/ledmatrix4j/font/Font9.java
@@ -1,0 +1,1005 @@
+package fr.craftinglabs.pi.ledmatrix4j.font;
+
+public class Font9 extends Font {
+
+    static public final Glyph A = new Glyph('A',
+            "00111000",
+            "01101100",
+            "11000110",
+            "11000110",
+            "11111110",
+            "11000110",
+            "11000110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph B = new Glyph('B',
+            "11111100",
+            "11000110",
+            "11000110",
+            "11111100",
+            "11000110",
+            "11000110",
+            "11111100",
+            "00000000",
+            "00000000");
+
+    static public final Glyph C = new Glyph('C',
+            "01111100",
+            "11000110",
+            "11000000",
+            "11000000",
+            "11000000",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+
+    static public final Glyph D = new Glyph('D',
+            "11111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11111100",
+            "00000000",
+            "00000000");
+
+    static public final Glyph E = new Glyph('E',
+            "11111110",
+            "11000000",
+            "11000000",
+            "11111100",
+            "11000000",
+            "11000000",
+            "11111110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph F = new Glyph('F',
+            "11111110",
+            "11000000",
+            "11000000",
+            "11111100",
+            "11000000",
+            "11000000",
+            "11000000",
+            "00000000",
+            "00000000");
+
+    static public final Glyph G = new Glyph('G',
+            "01111100",
+            "11000110",
+            "11000000",
+            "11001110",
+            "11000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+
+    static public final Glyph H = new Glyph('H',
+            "11000110",
+            "11000110",
+            "11000110",
+            "11111110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph I = new Glyph('I',
+            "1111110",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "1111110",
+            "0000000",
+            "0000000");
+
+    static public final Glyph J = new Glyph('J',
+            "00111110",
+            "00001100",
+            "00001100",
+            "00001100",
+            "00001100",
+            "11001100",
+            "01111000",
+            "00000000",
+            "00000000");
+
+    static public final Glyph K = new Glyph('K',
+            "11000110",
+            "11001100",
+            "11011000",
+            "11110000",
+            "11111000",
+            "11011100",
+            "11000110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph L = new Glyph('L',
+            "11000000",
+            "11000000",
+            "11000000",
+            "11000000",
+            "11000000",
+            "11000000",
+            "11111110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph M = new Glyph('M',
+            "10000010",
+            "11000110",
+            "11101110",
+            "11111110",
+            "11010110",
+            "11000110",
+            "11000110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph N = new Glyph('N',
+            "11000110",
+            "11100110",
+            "11110110",
+            "11011110",
+            "11001110",
+            "11000110",
+            "11000110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph O = new Glyph('O',
+            "01111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph P = new Glyph('P',
+            "11111100",
+            "11000110",
+            "11000110",
+            "11111100",
+            "11000000",
+            "11000000",
+            "11000000",
+            "00000000",
+            "00000000");
+    static public final Glyph Q = new Glyph('Q',
+            "01111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11011110",
+            "11001100",
+            "01110110",
+            "00000000",
+            "00000000");
+    static public final Glyph R = new Glyph('R',
+            "11111100",
+            "11000110",
+            "11000110",
+            "11111100",
+            "11011000",
+            "11001100",
+            "11000110",
+            "00000000",
+            "00000000");
+    static public final Glyph S = new Glyph('S', "01111100",
+            "11000110",
+            "11000000",
+            "01111100",
+            "00000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph T = new Glyph('T',
+            "1111110",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0000000",
+            "0000000");
+
+    static public final Glyph U = new Glyph('U',
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+
+    static public final Glyph V = new Glyph('V',
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01101100",
+            "00111000",
+            "00000000",
+            "00000000");
+
+    static public final Glyph W = new Glyph('W',
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11010110",
+            "11111110",
+            "01101100",
+            "00000000",
+            "00000000");
+    static public final Glyph X = new Glyph('X',
+            "11000110",
+            "01101100",
+            "00111000",
+            "00111000",
+            "01101100",
+            "11000110",
+            "10000010",
+            "00000000",
+            "00000000");
+    static public final Glyph Y = new Glyph('Y',
+            "1100110",
+            "1100110",
+            "1100110",
+            "0111100",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0000000",
+            "0000000");
+
+    static public final Glyph Z = new Glyph('Z',
+            "11111110",
+            "00001100",
+            "00011000",
+            "00110000",
+            "01100000",
+            "11000000",
+            "11111110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph a = new Glyph('a', "00000000",
+            "00000000",
+            "01111100",
+            "00000110",
+            "01111110",
+            "11000110",
+            "01111110",
+            "00000000",
+            "00000000");
+    static public final Glyph b = new Glyph('b',
+            "11000000",
+            "11000000",
+            "11111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11111100",
+            "00000000",
+            "00000000");
+    static public final Glyph c = new Glyph('c',
+            "00000000",
+            "00000000",
+            "01111100",
+            "11000110",
+            "11000000",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph d = new Glyph('d',
+            "00000110",
+            "00000110",
+            "01111110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111110",
+            "00000000",
+            "00000000");
+    static public final Glyph e = new Glyph('e',
+            "00000000",
+            "00000000",
+            "01111100",
+            "11000110",
+            "11111110",
+            "11000000",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph f = new Glyph('f', "001110",
+            "011000",
+            "111100",
+            "011000",
+            "011000",
+            "011000",
+            "011000",
+            "000000",
+            "000000");
+    static public final Glyph g = new Glyph('g',
+            "00000000",
+            "00000000",
+            "01111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111110",
+            "00000110",
+            "01111100");
+    static public final Glyph h = new Glyph('h',
+            "11000000",
+            "11000000",
+            "11111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "00000000",
+            "00000000");
+    static public final Glyph i = new Glyph('i',
+            "110",
+            "000",
+            "110",
+            "110",
+            "110",
+            "110",
+            "110",
+            "000",
+            "000");
+    static public final Glyph j = new Glyph('j',
+            "00000",
+            "00000",
+            "00110",
+            "00000",
+            "01110",
+            "00110",
+            "00110",
+            "00110",
+            "11100");
+    static public final Glyph k = new Glyph('k',
+            "1100000",
+            "1100000",
+            "1100110",
+            "1101100",
+            "1111000",
+            "1101100",
+            "1100110",
+            "0000000",
+            "0000000");
+    static public final Glyph l = new Glyph('l',
+            "110",
+            "110",
+            "110",
+            "110",
+            "110",
+            "110",
+            "110",
+            "000",
+            "000");
+    static public final Glyph m = new Glyph('m',
+            "00000000",
+            "00000000",
+            "11101100",
+            "11010110",
+            "11010110",
+            "11010110",
+            "11010110",
+            "00000000",
+            "00000000");
+    static public final Glyph n = new Glyph('n',
+            "00000000",
+            "00000000",
+            "11111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "00000000",
+            "00000000");
+    static public final Glyph o = new Glyph('o',
+            "00000000",
+            "00000000",
+            "01111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph p = new Glyph('p',
+            "00000000",
+            "00000000",
+            "11111100",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11111100",
+            "11000000",
+            "11000000");
+    static public final Glyph q = new Glyph('q',
+            "00000000",
+            "00000000",
+            "01111110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111110",
+            "00000110",
+            "00000110");
+    static public final Glyph r = new Glyph('r',
+            "000000",
+            "000000",
+            "110110",
+            "111000",
+            "110000",
+            "110000",
+            "110000",
+            "000000",
+            "000000");
+    static public final Glyph s = new Glyph('s',
+            "00000000",
+            "00000000",
+            "01111100",
+            "11000000",
+            "01111100",
+            "00000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph t = new Glyph('t',
+            "011000",
+            "011000",
+            "111100",
+            "011000",
+            "011000",
+            "011000",
+            "001110",
+            "000000",
+            "000000");
+    static public final Glyph u = new Glyph('u',
+            "00000000",
+            "00000000",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111110",
+            "00000000",
+            "00000000");
+    static public final Glyph v = new Glyph('v',
+            "00000000",
+            "00000000",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01101100",
+            "00111000",
+            "00000000",
+            "00000000");
+    static public final Glyph w = new Glyph('w',
+            "00000000",
+            "00000000",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11010110",
+            "01101100",
+            "00000000",
+            "00000000");
+    static public final Glyph x = new Glyph('x',
+            "00000000",
+            "00000000",
+            "11000110",
+            "01101100",
+            "00111000",
+            "01101100",
+            "11000110",
+            "00000000",
+            "00000000");
+    static public final Glyph y = new Glyph('y',
+            "00000000",
+            "00000000",
+            "11000110",
+            "11000110",
+            "11000110",
+            "11000110",
+            "01111110",
+            "00000110",
+            "01111100");
+    static public final Glyph z = new Glyph('z',
+            "00000000",
+            "00000000",
+            "11111110",
+            "00001100",
+            "00111000",
+            "01100000",
+            "11111110",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM0 = new Glyph('1',
+            "0111100",
+            "1100110",
+            "1100110",
+            "1100110",
+            "1100110",
+            "1100110",
+            "0111100",
+            "0000000",
+            "0000000");
+    static public final Glyph NUM1 = new Glyph('1',
+            "00110000",
+            "01110000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "0011000",
+            "1111110",
+            "0000000",
+            "0000000");
+    static public final Glyph NUM2 = new Glyph('2',
+            "01111100",
+            "11000110",
+            "00001100",
+            "00111000",
+            "01100000",
+            "11000000",
+            "11111110",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM3 = new Glyph('3',
+            "01111100",
+            "11000110",
+            "00000110",
+            "00011100",
+            "00000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM4 = new Glyph('4',
+            "00011100",
+            "00111100",
+            "01101100",
+            "11001100",
+            "11111110",
+            "00001100",
+            "00001100",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM5 = new Glyph('5',
+            "11111100",
+            "11000000",
+            "11111100",
+            "00000110",
+            "00000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM6 = new Glyph('6', "01111100",
+            "11000110",
+            "11000000",
+            "11111100",
+            "11000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM7 = new Glyph('7',
+            "11111110",
+            "00001100",
+            "00011000",
+            "00110000",
+            "00110000",
+            "00110000",
+            "00110000",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM8 = new Glyph('8',
+            "01111100",
+            "11000110",
+            "11000110",
+            "01111100",
+            "11000110",
+            "11000110",
+            "01111100",
+            "00000000",
+            "00000000");
+    static public final Glyph NUM9 = new Glyph('9',
+            "01111100",
+            "11000110",
+            "11000110",
+            "01111110",
+            "00000110",
+            "00001100",
+            "00111000",
+            "00000000",
+            "00000000");
+    static public final Glyph DOT = new Glyph('.',
+            "000",
+            "000",
+            "000",
+            "000",
+            "000",
+            "110",
+            "110",
+            "000",
+            "000");
+    static public final Glyph SLASH = new Glyph('/',
+            "000000110",
+            "000001100",
+            "000011000",
+            "000110000",
+            "001100000",
+            "011000000",
+            "110000000",
+            "000000000",
+            "000000000");
+
+    static public final Glyph SPACE = new Glyph(' ',
+            "00000",
+            "00000",
+            "00000",
+            "00000",
+            "00000",
+            "00000",
+            "00000",
+            "00000",
+            "00000");
+    static public final Glyph BACKSLASH = new Glyph('\\',
+            "1100000 0",
+            "011000000",
+            "001100000",
+            "000110000",
+            "000011000",
+            "000001100",
+            "000000110",
+            "000000000",
+            "000000000");
+    static public final Glyph DASH = new Glyph('-',
+            "00000",
+            "00000",
+            "00000",
+            "11110",
+            "00000",
+            "00000",
+            "00000",
+            "00000",
+            "00000");
+    static public final Glyph BRACKET_CLOSE = new Glyph(']',
+            "1110",
+            "0110",
+            "0110",
+            "0110",
+            "0110",
+            "0110",
+            "1110",
+            "0000",
+            "0000");
+    static public final Glyph BRACKET_OPEN = new Glyph('[', "1110",
+            "1100",
+            "1100",
+            "1100",
+            "1100",
+            "1100",
+            "1110",
+            "0000",
+            "0000");
+    static public final Glyph CURLY_OPEN = new Glyph('{',
+            "0001110",
+            "0011000",
+            "0011000",
+            "1110000",
+            "0011000",
+            "0011000",
+            "0001110",
+            "0000000",
+            "0000000");
+    static public final Glyph CURLY_CLOSE = new Glyph('}',
+            "1110000",
+            "0011000",
+            "0011000",
+            "0001110",
+            "0011000",
+            "0011000",
+            "1110000",
+            "0000000",
+            "0000000");
+    static public final Glyph TILDE = new Glyph('~',
+            "00000000",
+            "00000000",
+            "00000000",
+            "01110110",
+            "11011100",
+            "00000000",
+            "00000000",
+            "00000000",
+            "00000000");
+    static public final Glyph UNDERSCORE = new Glyph('_',
+            "000000",
+            "000000",
+            "000000",
+            "000000",
+            "000000",
+            "000000",
+            "111110",
+            "000000",
+            "000000");
+    static public final Glyph COMMA = new Glyph(',',
+            "0000",
+            "0000",
+            "0000",
+            "0000",
+            "0000",
+            "0110",
+            "0110",
+            "1100",
+            "0000");
+    static public final Glyph SEMI_COLON = new Glyph(';',
+            "000",
+            "110",
+            "110",
+            "000",
+            "110",
+            "110",
+            "010",
+            "000",
+            "000");
+    static public final Glyph COLON = new Glyph(':',
+            "000",
+            "110",
+            "110",
+            "000",
+            "110",
+            "110",
+            "000",
+            "000",
+            "000");
+    static public final Glyph EXCLAMATION = new Glyph('!',
+            "110",
+            "110",
+            "110",
+            "110",
+            "110",
+            "000",
+            "110",
+            "000",
+            "000");
+
+    static public final Glyph QUESTION = new Glyph('?',
+            "0111100",
+            "1100110",
+            "0001100",
+            "0011000",
+            "0010000",
+            "0000000",
+            "0011000",
+            "0000000",
+            "0000000");
+
+    static public final Glyph AROBASE = new Glyph('@',
+            "01111100",
+            "10000010",
+            "10111010",
+            "10101010",
+            "10110110",
+            "10000000",
+            "01111110",
+            "00000000",
+            "00000000");
+
+    static public final Glyph SHARP = new Glyph('#',
+            "01101100",
+            "01101100",
+            "11111110",
+            "01101100",
+            "11111110",
+            "01101100",
+            "01101100",
+            "00000000",
+            "00000000");
+    static public final Glyph APOSTROPHE = new Glyph('\'',
+            "110",
+            "110",
+            "010",
+            "000",
+            "000",
+            "0000",
+            "000",
+            "000",
+            "000");
+    static public final Glyph QUOTATION = new Glyph('\"',
+            "110110",
+            "110110",
+            "010010",
+            "000000",
+            "000000",
+            "000000",
+            "000000",
+            "000000",
+            "000000");
+    static public final Glyph PERCENT = new Glyph('%',
+            "0000000",
+            "0000000",
+            "1100110",
+            "1101100",
+            "0011000",
+            "0110110",
+            "1100110",
+            "0000000",
+            "0000000");
+    static public final Glyph AMPERSAND = new Glyph('&',
+            "011000",
+            "100100",
+            "101000",
+            "010000",
+            "101010",
+            "100100",
+            "011010",
+            "000000",
+            "000000");
+    static public final Glyph ASTERISK = new Glyph('*',
+            "0000",
+            "0000",
+            "0000",
+            "0000",
+            "1010",
+            "0100",
+            "1010",
+            "0000",
+            "0000");
+    static public final Glyph PARENTHESE_OPEN = new Glyph('(',
+            "00110",
+            "01100",
+            "11000",
+            "11000",
+            "11000",
+            "01100",
+            "00110",
+            "00000",
+            "00000");
+    static public final Glyph PARENTHESE_CLOSE = new Glyph(')',
+            "11000",
+            "01100",
+            "00110",
+            "00110",
+            "00110",
+            "01100",
+            "11000",
+            "00000",
+            "00000");
+    static public final Glyph DOLLAR = new Glyph('$',
+            "00100",
+            "01110",
+            "10000",
+            "11100",
+            "00010",
+            "11100",
+            "00100",
+            "00000",
+            "00000");
+    static public final Glyph EURO = new Glyph('€',
+            "0011110",
+            "0100000",
+            "1111000",
+            "0100000",
+            "1111000",
+            "0100000",
+            "0011110",
+            "0000000",
+            "0000000");
+
+    static {
+            add(Font9.A);
+        add(Font9.B);
+        add(Font9.C);
+        add(Font9.D);
+        add(Font9.E);
+        add(Font9.F);
+        add(Font9.G);
+        add(Font9.H);
+        add(Font9.I);
+        add(Font9.J);
+        add(Font9.K);
+        add(Font9.L);
+        add(Font9.M);
+        add(Font9.N);
+        add(Font9.O);
+        add(Font9.P);
+        add(Font9.Q);
+        add(Font9.R);
+        add(Font9.S);
+        add(Font9.T);
+        add(Font9.U);
+        add(Font9.V);
+        add(Font9.W);
+        add(Font9.X);
+        add(Font9.Y);
+        add(Font9.Z);
+        add(Font9.a);
+        add(Font9.b);
+        add(Font9.c);
+        add(Font9.d);
+        add(Font9.e);
+        add(Font9.f);
+        add(Font9.g);
+        add(Font9.h);
+        add(Font9.i);
+        add(Font9.j);
+        add(Font9.k);
+        add(Font9.l);
+        add(Font9.m);
+        add(Font9.n);
+        add(Font9.o);
+        add(Font9.p);
+        add(Font9.q);
+        add(Font9.r);
+        add(Font9.s);
+        add(Font9.t);
+        add(Font9.u);
+        add(Font9.v);
+        add(Font9.w);
+        add(Font9.x);
+        add(Font9.y);
+        add(Font9.z);
+        add(Font9.NUM0);
+        add(Font9.NUM1);
+        add(Font9.NUM2);
+        add(Font9.NUM3);
+        add(Font9.NUM4);
+        add(Font9.NUM5);
+        add(Font9.NUM6);
+        add(Font9.NUM7);
+        add(Font9.NUM8);
+        add(Font9.NUM9);
+        add(Font9.DASH);
+        add(Font9.SLASH);
+        add(Font9.DOT);
+        add(Font9.COLON);
+        add(Font9.COMMA);
+        add(Font9.SPACE);
+        add(Font9.QUESTION);
+        add(Font9.EXCLAMATION);
+        add(Font9.BACKSLASH);
+        add(Font9.AMPERSAND);
+        add(Font9.APOSTROPHE);
+        add(Font9.AROBASE);
+        add(Font9.ASTERISK);
+        add(Font9.BRACKET_CLOSE);
+        add(Font9.BRACKET_OPEN);
+        add(Font9.CURLY_OPEN);
+        add(Font9.CURLY_CLOSE);
+        add(Font9.DOLLAR);
+        add(Font9.TILDE);
+        add(Font9.UNDERSCORE);
+        add(Font9.SEMI_COLON);
+        add(Font9.SHARP);
+        add(Font9.QUOTATION);
+        add(Font9.PERCENT);
+        add(Font9.PARENTHESE_CLOSE);
+        add(Font9.PARENTHESE_OPEN);
+        add(Font9.EURO);
+    }
+}
