+++ ledmatrix4j/src/main/java/fr/craftinglabs/pi/ledmatrix4j/io/MatrixIO.java
@@ -1,0 +1,12 @@
+package fr.craftinglabs.pi.ledmatrix4j.io;
+
+public interface MatrixIO {
+
+    void incrementCounter();
+
+    void resetCounter();
+
+    void releaseShift() ;
+
+    void shiftLine(byte value);
+}
